Working on commenting the managers
