+++ Assets/Scripts/Environment/Lights/LightFlash.cs
@@ -1,20 +1,41 @@
 using UnityEngine;
 
+/// <summary>
+/// Flash a object with the specified on and off times.
+/// </summary>
 public class LightFlash : MonoBehaviour
 {
+    /// <summary>
+    /// The desired object to flash.
+    /// </summary>
     [SerializeField]
     private GameObject objectToFlash;
 
+    /// <summary>
+    /// How long to show the object in seconds.
+    /// </summary>
     [SerializeField]
     private float onTime = 0.3f;
 
+    /// <summary>
+    /// How long to hide the object in seconds.
+    /// </summary>
     [SerializeField]
     private float offTime = 0.5f;
 
+    /// <summary>
+    /// The current time for the current state.
+    /// </summary>
     private float currentTime = 0;
 
+    /// <summary>
+    /// Is this flash currently flashing.
+    /// </summary>
     private bool isActive = false;
 
+    /// <summary>
+    /// The flash state. True = currently on.
+    /// </summary>
     private bool isFlashState = false;
 
 
@@ -42,18 +63,27 @@ public class LightFlash : MonoBehaviour
         }
     }
 
+    /// <summary>
+    /// Begin the flashing process.
+    /// </summary>
     public void StartFlashing()
     {
         isActive = true;
         TurnOn();
     }
 
+    /// <summary>
+    /// Stop the flashing process.
+    /// </summary>
     public void StopFlashing()
     {
         isActive = false;
         TurnOff();
     }
 
+    /// <summary>
+    /// Set the states for on.
+    /// </summary>
     private void TurnOn()
     {
         isFlashState = true;
@@ -61,6 +91,9 @@ public class LightFlash : MonoBehaviour
         objectToFlash.SetActive(isFlashState);
     }
 
+    /// <summary>
+    /// Set the states for off.
+    /// </summary>
     private void TurnOff()
     {
         isFlashState = false;
