+++ Assets/Scripts/Environment/Projectiles/ProjectileScript.cs
@@ -1,18 +1,24 @@
-using TMPro;
 using UnityEngine;
 
-// By Vincent Pressey
-
+/// <summary>
+/// A moving physics object that will deal damage if it hits an object that is damageable, this will also destroy the object on collision. 
+/// </summary>
 public class ProjectileScript : MonoBehaviour
 {
     /// <summary>
-    /// Damage dealt to object when projectile makes contact
+    /// Damage dealt to object when projectile makes contact.
     /// </summary>
     [HideInInspector]
     public float ProjectileDamage = 12;
 
+    /// <summary>
+    /// Did the projectile already hit something. Prevents entities getting hit multiple times.
+    /// </summary>
     private bool hasHit;
 
+    /// <summary>
+    /// The source of the projectile, ideally the entity that fired it.
+    /// </summary>
     public Transform SourceForProjectile;
 
     private void OnTriggerEnter(Collider collider)
