+++ Assets/Scripts/Interfaces/IMeleeable.cs
@@ -1,6 +1,12 @@
 using UnityEngine;
 
+/// <summary>
+/// Allows the object to be melee-ed by the player.
+/// </summary>
 public interface IMeleeable
 {
+    /// <summary>
+    /// Melee this object.
+    /// </summary>
     public void MeleeObject();
 }
