+++ Assets/Scripts/Manager/Level.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 43d7a693580ab26468969fe7d43877b0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
