+++ Assets/Scripts/Manager/RunBased.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c63030bf61730d42b4ed0c3a30f6b33
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
